[reformat] with prettier
