+++ src/components/FilterBar/FilterBar.js
@@ -1,0 +1,15 @@
+import React, { PureComponent } from 'react';
+
+import './styles.css';
+
+class FilterBar extends PureComponent {
+  render() {
+    return (
+      <div className="filter-bar">
+        <h1>Search Filters</h1>
+      </div>
+    );
+  }
+}
+
+export default FilterBar;
