+++ src/components/FilterBar/index.js
@@ -1,0 +1,3 @@
+import FilterBar from './FilterBar';
+
+export default FilterBar;
