--- src/components/View-btn-group/index.js
@@ -1,3 +1,0 @@
-import ViewBtnGroup from './btn-group';
-
-export default ViewBtnGroup;
