+++ src/components/ViewBtnGroup/index.js
@@ -1,0 +1,3 @@
+import ViewBtnGroup from './BtnGroup';
+
+export default ViewBtnGroup;
