--- src/components/filter-bar/index.js
@@ -1,4 +1,0 @@
-import FilterBar from './filter-bar';
-
-
-export default FilterBar;
